Update and rename bash.sh to zsh.sh
Converting to zsh since it is now the default on MACOS.  Need to check if there is anything else needed in zshenv or if zlogin/zprofile is a better alternative.  Also needs to be tested in Ubuntu.  Do we still need to increase the watcher limit in Ubuntu?  I believe this was added b/c of issues with using Atom as the commit message editor. (?)